Script updates due to site destination name change (#40)

+++ tool/check-links.sh
@@ -5,19 +5,23 @@ set -e -o pipefail
 [[ -z "$DART_SITE_ENV_DEFS" ]] && . ./scripts/env-set.sh
 
 PORT=${SITE_LOCALHOST_PORT:-5000}
-CHECK_EXIT_CODE=0
 
-if [ ! -e "publish" ]; then
-  echo "Site not built, skipping link check."
-  exit $CHECK_EXIT_CODE
+set +e
+SITE=$(grep '^destination:' _config.yml | awk '{ print $2}')
+set -e
+: ${SITE:-_site}
+
+if [ ! -e "./$SITE" ]; then
+  echo "INFO: $SITE directory not found. Site not built? Skipping link checks."
+  exit 0
 fi
 
-set -x
-((superstatic --port $PORT > /dev/null  2>&1) \
-  || echo "Failed to launch superstatic server. Maybe it is already running?") &
-SERVER_PID=$!
-
-sleep 4
+if (set -x; superstatic --port $PORT > /dev/null 2>&1) then
+  SERVER_PID=$!
+  sleep 4
+else
+  echo "WARNING: Failed to launch superstatic server. I'll assume it is already running."
+fi
 
 # Don't check for external links yet since it seems to cause problems on Travis: --external
 pub run linkcheck \
@@ -29,8 +33,10 @@ set +x
 
 if ! grep '^\s*0 errors' $TMP/linkcheck-log.txt | wc -l > /dev/null; then
   CHECK_EXIT_CODE=1
+else
+  CHECK_EXIT_CODE=0
 fi
 
-kill $SERVER_PID
+if [[ -n $SERVER_PID ]]; then kill $SERVER_PID; fi
 
 exit $CHECK_EXIT_CODE
